Add BringerOfDeath boss and multiple enemy fixes
Implement BringerOfDeath boss (entity/BringerOfDeath.java) with custom hitbox, draw/flip logic, behaviors, special attacks and hurt/death effects. Add skullwolf.png resource. Update EnemyManager to load bringer sprites and added helper loader. Update LoadSave to parse map ID for bringer and spawn it. Add constants for bringer states, sprite sizes, exp and HP in Konstanta. Fix BlueGolem draw/attack box, animation bounds and multi-hit; fix Slime attack box positioning and attack timing; change attackDistance types/values in Enemy/DemonBoss. Misc robustness and debug-draw improvements.

+++ src/entity/DemonBoss.java
@@ -27,7 +27,7 @@ public class DemonBoss extends Boss {
         initAttackBox();
 //      walkSpeed = 0.22f * GameCore.SCALE;
         walkSpeed = 0.73f * GameCore.SCALE;
-        attackDistance = GameCore.TILES_SIZE * 2f;
+        attackDistance = (int) (GameCore.TILES_SIZE * 2f);
 //      aniSpeed = 12;
         aniSpeed = 4;
     }
